Extract confirmation_flow.rs from turn.rs (parent #680)
Create new sibling module `src/agent/loop_run/confirmation_flow.rs`
to host the Issue #576 / #579 / #580 confirmation-flow plumbing
(preflight skip-reason predicates, outcome → parse-status projection,
log-payload wrappers, per-turn memo cache lookup). Small follow-on
extraction in the meta-issue #680 reduction effort.

Migrated to `confirmation_flow.rs` (`pub(super)`):

* `should_writeback_first_pass(work_mode_confirm_called_this_turn)
    -> bool` (CB-001 / Issue #576)
* `effective_turn_index_for_stage(stage_label, current_turn_index)
    -> usize` (CB-004 / Issue #576)
* `preflight_work_mode_skip_reason(...) -> Option<WorkModeSkipReason>`
* `preflight_feedback_kind_skip_reason(...) -> Option<FeedbackKindSkipReason>`
* `preflight_quality_confirm_skip_reason(...) -> Option<QualityConfirmSkipReason>`
* `quality_confirm_cached_result(last_result, content_hash)
    -> Option<QualityConfirmation>` (Issue #580 per-turn memo)
* `work_mode_confirm_parse_status(&outcome) -> ParseStatus` (outcome
    → enum projection)
* `log_work_mode_confirm_outcome(...)` (wraps `build_work_mode_confirm_log_payload`)
* `log_feedback_kind_confirm_outcome(...)` (wraps `build_feedback_kind_confirm_log_payload`)
* `log_quality_confirm_outcome(...)` (wraps `build_quality_confirm_log_payload`)
* `override_feedback_kind_from_outcome(&outcome, &first_pass)
    -> Option<FeedbackKind>` (SecondPassOverridden extractor)

`loop_run.rs` adjustments:

* Registered `mod confirmation_flow;` with a parent-#680 doc comment.

`turn.rs` adjustments:

* Removed the 11 fn definitions (originally at lines 813-993).
* Added 10 imports via new `use super::confirmation_flow::{...}` block.
* Pruned now-unused symbols from existing import blocks:
  - `super::work_mode_confirm::{...}`: removed
    `ParseStatus as WorkModeConfirmParseStatus`,
    `build_work_mode_confirm_log_payload`.
  - `super::feedback_kind_confirm::{...}`: removed
    `build_feedback_kind_confirm_log_payload`.
  - `super::quality_confirm::{...}`: removed
    `build_quality_confirm_log_payload`.
* Test mod ref `super::work_mode_confirm_parse_status` rewrote to
  `super::super::confirmation_flow::work_mode_confirm_parse_status`.

LOC delta:

* `turn.rs`: 28,233 → 28,043 (-190 LOC)
* `confirmation_flow.rs`: 0 → 222 (new)
* Cumulative `turn.rs`: 39,489 → 28,043 (-11,446 LOC, -29.0%)

`pub(super)` limited / no facade re-export (DR3-001).
`turn.rs` remains the only in-crate consumer.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
